+++ TimePicker/TimePicker.js
@@ -73,13 +73,14 @@ const indexOfMeridiem = (time, meridiems) => {
 };
 
 // Label formatter
-const labelFormatter = new DateFmt({
+let formatter;
+const labelFormatter = () => formatter || (formatter =  new DateFmt({
 	type: 'time',
 	useNative: false,
 	timezone: 'local',
 	length: 'full',
 	date: 'dmwy'
-});
+}));
 
 const dateTimeConfig = {
 	customProps: function (i18n, value, {meridiemLabel}) {
@@ -181,13 +182,13 @@ const dateTimeConfig = {
 		const meridiemEnabled = clockPref === '12';
 
 		const filter = meridiemEnabled ? includeMeridiem : excludeMeridiem;
-		const order = labelFormatter.getTemplate()
+		const order = labelFormatter().getTemplate()
 			.replace(/'.*?'/g, '')
 			.match(filter)
 			.map(s => s[0].toLowerCase());
 
 		return {
-			formatter: labelFormatter,
+			formatter: labelFormatter(),
 			meridiemEnabled,
 			meridiemLabels,
 			meridiemRanges,
@@ -257,7 +258,7 @@ const timeToLocaleString = (time) => {
 		return null;
 	}
 
-	return labelFormatter.format(time);
+	return labelFormatter().format(time);
 };
 
 export default TimePicker;